refactor(mobile): extract image viewer actions to protocol
- Create ImageCarouselViewControllerProtocol to define image interaction methods
- Move save, copy, and share image methods from ImageViewerController to protocol implementation
- Remove setupOptionsButton method from ImageViewerController
- Add options menu directly in ImageCarouselViewController
- Enable page control user interaction
- Simplify image action handling across image viewer components
