+++ apps/mobile/native/ios/Packages/ImageViewer_swift/ImageCarouselViewControllerProtocol.swift
@@ -1,0 +1,15 @@
+//
+//  ImageCarouselViewControllerProtocol.swift
+//  FollowNative
+//
+//  Created by Innei on 2025/3/12.
+//
+
+import UIKit
+
+protocol ImageCarouselViewControllerProtocol {
+  func saveImageToPhotos()
+  func copyImageToClipboard()
+  func shareImage()
+
+}
